requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-filter==23.5
idna==3.6
requests==2.31.0
signup==0.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
